custom_user/cust_user/requirements.txt
Django==1.11
django-crispy-forms
djangorestframework
Pillow==3.1.0
urllib3

